requirements.txt
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.52
python-dotenv==1.0.1
requests==2.31.0
flask==3.0.2
mistralai==1.0.2
langchain-mistralai==0.1.1